@@ -3,15 +3,15 @@ package com.kizitonwose.colorpreference;
 import android.app.Activity;
 import android.content.Context;
 import android.content.ContextWrapper;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 /**
  * @author Kizito Nwose
  */
 
-public class Utils {
+class Utils {
     @Nullable
-    public static Activity resolveContext(Context context) {
+    static Activity resolveContext(Context context) {
         if (context instanceof Activity) {
             return (Activity) context;
         } else if (context instanceof ContextWrapper) {
